Créé un store pinia pour gérer le thème
