--- client/composables/useDsfrScheme.ts
@@ -1,7 +1,0 @@
-export function useDsfrScheme () {
-  const isThemeModalOpen = useState('theme-modal', () => false)
-
-  return {
-    isThemeModalOpen
-  }
-}
